backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9

# Haystack 2.0
haystack-ai>=2.11.0
farm-haystack>=2.11.0

# Vector Stores & Search
qdrant-client>=1.12.0
elasticsearch>=8.15.0

# Embeddings & LLMs
mistralai>=1.0.0
openai>=1.50.0
anthropic>=0.40.0

# CrewAI
crewai>=0.80.0

# Document Processing
pypdf>=5.1.0
python-docx>=1.1.0
python-pptx>=1.0.0
unstructured>=0.15.0
pillow>=11.0.0

# Celery for Background Tasks
celery>=5.4.0
redis>=5.2.0
django-redis>=5.4.0

# Testing
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0
factory-boy>=3.3.0
faker>=30.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
